CB-493 style: modify add pet button
반려동물 등록 버튼 스타일 수정
- 마이페이지 버튼 헤더, 리스트 버튼 분리
