+++ index.html
@@ -13,11 +13,9 @@
     <div id="confirm"></div>
     <div id="select"></div>
     <script type="module" src="/src/main.tsx"></script>
-    <link
-      href="https://fonts.googleapis.com/css?family=Roboto:100,300,400,500,700"
-      rel="stylesheet"
-    />
+    <link href="https://fonts.googleapis.com/css?family=Roboto:400,500,600,700" rel="stylesheet" />
     <script
+      async
       type="text/javascript"
       src="//dapi.kakao.com/v2/maps/sdk.js?appkey=<%= KAKAO_KEY %>&libraries=services,clusterer"
     ></script>
