+++ src/styles/index.css
@@ -1,1 +1,1 @@
-@import url(https://fonts.googleapis.com/css2?family=IBM+Plex+Sans+KR:wght@400;600&family=Noto+Sans+KR:wght@300;400;500;700&display=swap);
+@import url(https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@400;500;600;700&display=swap);
